feat(#97): Integration of JMH for Benchmarking Performance #97
