+++ .gitignore
@@ -45,3 +45,9 @@ bin/
 gradle.properties
 !gradle.properties
 !**/src/main/**/gradle.properties
+
+### Images
+*.jpg
+*.jpeg
+*.png
+/benchmark-data/augmentation-benchmark-images/*
